+++ mmaction/models/heads/roi_head.py
@@ -1,8 +1,7 @@
-import warnings
-
 import numpy as np
 
 from mmaction.core.bbox import bbox2result
+from mmaction.utils import import_module_error_class
 
 try:
     from mmdet.core.bbox import bbox2roi
@@ -10,8 +9,6 @@ try:
     from mmdet.models.roi_heads import StandardRoIHead
     mmdet_imported = True
 except (ImportError, ModuleNotFoundError):
-    warnings.warn('Please install mmdet to use bbox2roi, MMDET_HEADS '
-                  'and StandardRoIHead')
     mmdet_imported = False
 
 if mmdet_imported:
@@ -86,5 +83,6 @@ if mmdet_imported:
             return det_bboxes, det_labels
 else:
     # Just define an empty class, so that __init__ can import it.
+    @import_module_error_class('mmdet')
     class AVARoIHead:
         pass